[Runner] Clarify message printed when killing a process
It previously said it was going to SIGTERM the process after a few
milliseconds, if that process had already exited.

Now it only kills processes that do *not* die after receiving a SIGTERM.

+++ backend/shared/src/main/java/de/aaaaaaah/velcom/runner/shared/ProgramExecutor.java
@@ -80,11 +80,19 @@ public class ProgramExecutor {
 				processHandle.destroy();
 				// Allow time for graceful shutdown
 				try {
+					LOGGER.info(
+						"Waiting for {} millis for the thread to die",
+						timeToForceKillMillis
+					);
 					process.waitFor(timeToForceKillMillis, TimeUnit.MILLISECONDS);
 				} catch (InterruptedException ignored) {
 				}
-				LOGGER.info("Nuking uncooperative process (SIGKILL)");
-				processHandle.destroyForcibly();
+				if (processHandle.isAlive()) {
+					LOGGER.info("Nuking uncooperative process (SIGKILL)");
+					processHandle.destroyForcibly();
+				} else {
+					LOGGER.info("Process was nice and died properly!");
+				}
 				collectedException.set(new CancellationException("Killed process!"));
 			}
 			try {
